added new set of slides
Added the How to report a power analysis slide set. As I had problems with formatting when using all the pictures and texts from the original slide deck I just added the examples as screenshots. If the quality of the pictures is too low let me know and ill figure out how to properly format them